# docs/deliverables/system_docs_20260729/star-rail-agents.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Star Rail Agents",
    "description": "基于 Multi-Agent 与 RAG 的游戏智能助手平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/catalog/characters": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Characters",
        "operationId": "list_characters_api_v1_catalog_characters_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "element",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Element"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Path"
            }
          },
          {
            "name": "rarity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 4
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rarity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/characters/{character_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Character",
        "operationId": "get_character_api_v1_catalog_characters__character_id__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/characters/{character_id}/progression": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Character Progression",
        "operationId": "get_character_progression_api_v1_catalog_characters__character_id__progression_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/characters/{character_id}/progression/calculate": {
      "post": {
        "tags": [
          "catalog"
        ],
        "summary": "Calculate Character Progression",
        "operationId": "calculate_character_progression_api_v1_catalog_characters__character_id__progression_calculate_post",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Character Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressionCalculateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionCalculateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/items": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Items",
        "operationId": "list_items_api_v1_catalog_items_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 2000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/items/{item_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Item",
        "operationId": "get_item_api_v1_catalog_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/lightcones/{entry_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Lightcone",
        "operationId": "get_lightcone_api_v1_catalog_lightcones__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LightconeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/relics/{entry_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Relic",
        "operationId": "get_relic_api_v1_catalog_relics__entry_id__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelicDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/{kind}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "List Knowledge Entities",
        "operationId": "list_knowledge_entities_api_v1_catalog__kind__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntityListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/{kind}/{entry_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Knowledge Entity",
        "operationId": "get_knowledge_entity_api_v1_catalog__kind___entry_id__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntityDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rarity": {
            "type": "integer",
            "title": "Rarity"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "element": {
            "type": "string",
            "title": "Element"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "energy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Energy"
          },
          "stats": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Stats"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/NamedDescription"
            },
            "type": "array",
            "title": "Skills"
          },
          "eidolons": {
            "items": {
              "$ref": "#/components/schemas/NamedDescription"
            },
            "type": "array",
            "title": "Eidolons"
          },
          "related_items": {
            "items": {
              "$ref": "#/components/schemas/RelatedItem"
            },
            "type": "array",
            "title": "Related Items"
          },
          "recommended_lightcones": {
            "items": {
              "$ref": "#/components/schemas/EntityRecommendation"
            },
            "type": "array",
            "title": "Recommended Lightcones"
          },
          "recommended_relics": {
            "items": {
              "$ref": "#/components/schemas/EntityRecommendation"
            },
            "type": "array",
            "title": "Recommended Relics"
          },
          "story": {
            "type": "string",
            "title": "Story",
            "default": ""
          },
          "community_source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community Source Url"
          },
          "portrait_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portrait Url"
          },
          "element_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Element Icon Url"
          },
          "path_icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Icon Url"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "rarity",
          "path",
          "element",
          "source_url"
        ],
        "title": "CharacterDetail"
      },
      "CharacterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CharacterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "CharacterListResponse"
      },
...
      "CharacterSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rarity": {
            "type": "integer",
            "title": "Rarity"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "element": {
            "type": "string",
            "title": "Element"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "rarity",
          "path",
          "element"
        ],
        "title": "CharacterSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "background": {
            "type": "string",
            "title": "Background",
            "default": ""
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "pile_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pile Limit"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "related_entities": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__catalog__RelatedEntity"
            },
            "type": "array",
            "title": "Related Entities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "rarity",
          "type"
        ],
        "title": "ItemDetail"
      },
      "ItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RelatedItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ItemListResponse"
      },
...
      "KnowledgeEntityListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEntitySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "KnowledgeEntityListResponse"
      },
...
      "LightconeDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle",
            "default": ""
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "image_status": {
            "type": "string",
            "title": "Image Status",
            "default": "official"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "set_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Type"
          },
          "piece_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Piece Count"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/NamedDescription"
            },
            "type": "array",
            "title": "Sections"
          },
          "related_entities": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__catalog__RelatedEntity"
            },
            "type": "array",
            "title": "Related Entities"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "lore_description": {
            "type": "string",
            "title": "Lore Description",
            "default": ""
          },
          "effect_name": {
            "type": "string",
            "title": "Effect Name",
            "default": ""
          },
          "superimposition_effects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Superimposition Effects"
          },
          "level_stats": {
            "items": {
              "$ref": "#/components/schemas/LightconeLevelStats"
            },
            "type": "array",
            "title": "Level Stats"
          },
          "promotion_materials": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/LightconeMaterial"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Promotion Materials"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "LightconeDetail"
      },
...
      "ProgressionCalculateRequest": {
        "properties": {
          "from_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "From Level"
          },
          "to_level": {
            "type": "integer",
            "maximum": 80.0,
            "minimum": 1.0,
            "title": "To Level"
          },
          "skill_ranges": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SkillRange"
            },
            "type": "object",
            "title": "Skill Ranges"
          }
        },
        "type": "object",
        "required": [
          "from_level",
          "to_level"
        ],
        "title": "ProgressionCalculateRequest"
      },
      "ProgressionCalculateResponse": {
        "properties": {
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "archetype": {
            "type": "string",
            "title": "Archetype"
          },
          "from_level": {
            "type": "integer",
            "title": "From Level"
          },
          "to_level": {
            "type": "integer",
            "title": "To Level"
          },
          "skill_ranges": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SkillRange"
            },
            "type": "object",
            "title": "Skill Ranges"
          },
          "total_by_key": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Total By Key"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/ProgressionMaterialResponse"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "archetype",
          "from_level",
          "to_level",
          "skill_ranges",
          "total_by_key",
          "materials"
        ],
        "title": "ProgressionCalculateResponse"
      },
...
      "ProgressionProfile": {
        "properties": {
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "archetype": {
            "type": "string",
            "title": "Archetype"
          },
          "tracks": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ProgressionTrack"
            },
            "type": "object",
            "title": "Tracks"
          },
          "ascension_gates": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Ascension Gates"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "archetype",
          "tracks",
          "ascension_gates"
        ],
        "title": "ProgressionProfile"
      },
      "ProgressionTrack": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "max_level": {
            "type": "integer",
            "title": "Max Level"
          }
        },
        "type": "object",
        "required": [
          "label",
          "max_level"
        ],
        "title": "ProgressionTrack"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地 FastAPI"
    }
  ]
}